#21 Fix e-notation
